Initial draft of HTTP in and out
Simple MVC endpoint: /streams/{path} where {path} is the name
of the MessageChannel.

Matches input with output if use enables binding on a Processor
interface (or any interface with precisley one input and one
output). Otherwise inputs and outputs are disconnected - you
POST to an input and GET from an output.

Missing:

* No headers, payloads only
* GET from /output buffers items and only returns them once