Support for POJOs coming in as JSON
As a special case, we can push the conversion of JSON POJOs down
to the Spring Cloud Stream layer (instead of asking Spring MVC
to do it). It might actually be better to do something similar
for other incoming mime types - the problem is that Stream knows
about the target data types alayer below MVC. Alternatively
we could switch to custom MVC handlers that delegate somehow
to the Stream converters.