Use MessagingTemplate to send messages
In principle we might be able to send-and-receive to a Processor
to capture the responses synchronously if they are available.
It might even prove possible for the user to opt into synchronous
responses through a request parameter.

This attempt doesn't work because the messaging template waits for
a response on a temporary reply queue, and Spring Cloud Stream
ignores that (so the tests stall and wait forever for a response
that went to a different channel).

Make it work by hacking in a copy of StreamListenerMessageHandler